Trial classes and Experiment subclasses
- Changed Experiment to an abstract class
      + Implementing addTrial(Trial trial)
      + Added Binomial, Count, Measurement, and NonNegative subclasses of Experiment
- Added Trial class
      + Added BinomialTrial, CountTrial, MeasurementTrial, and NonNegativeTrial to represent trial for each type of experiment
- Adjusted ExperimentManager to represent the changes
- Added the experiment package

TESTING REQUIRED
