+++ app/src/main/java/com/example/crowdtest/experiments/BinomialTrial.java
@@ -1,4 +1,4 @@
-package com.example.crowdtest;
+package com.example.crowdtest.experiments;
 
 public class BinomialTrial extends Trial {
     private boolean success;
