+++ app/src/main/java/com/example/crowdtest/experiments/Count.java
@@ -1,10 +1,12 @@
-package com.example.crowdtest;
+package com.example.crowdtest.experiments;
+
+import com.example.crowdtest.Experimenter;
 
 import java.util.ArrayList;
 import java.util.Collection;
 
-public class Binomial extends Experiment {
-    private Collection <BinomialTrial> trials;
+public class Count extends Experiment {
+    private Collection <CountTrial> trials;
 
     /**
      * Experiment constructor
@@ -12,13 +14,13 @@ public class Binomial extends Experiment {
      * @param owner         Owner of the experiment
      * @param experimentID  A unique ID for this experiment
      */
-    public Binomial(Experimenter owner, String experimentID) {
+    public Count(Experimenter owner, String experimentID) {
         super(owner, experimentID);
         trials = new ArrayList<>();
     }
 
     @Override
     public void addTrial(Trial trial) {
-        trials.add((BinomialTrial) trial);
+        trials.add((CountTrial) trial);
     }
 }
