+++ app/src/main/java/com/example/crowdtest/experiments/Experiment.java
@@ -1,4 +1,7 @@
-package com.example.crowdtest;
+package com.example.crowdtest.experiments;
+
+import com.example.crowdtest.Experimenter;
+import com.example.crowdtest.Question;
 
 import java.util.ArrayList;
