+++ app/src/main/java/com/example/crowdtest/experiments/MeasurementTrial.java
@@ -1,4 +1,4 @@
-package com.example.crowdtest;
+package com.example.crowdtest.experiments;
 
 public class MeasurementTrial extends Trial {
     private double measurement;
