+++ app/src/main/java/com/example/crowdtest/experiments/NonNegativeTrial.java
@@ -1,4 +1,4 @@
-package com.example.crowdtest;
+package com.example.crowdtest.experiments;
 
 public class NonNegativeTrial extends Trial {
     private int count;
