+++ app/src/main/java/com/example/crowdtest/experiments/Trial.java
@@ -1,4 +1,4 @@
-package com.example.crowdtest;
+package com.example.crowdtest.experiments;
 
 import android.location.Location;
